requirements.txt
torch
torchvision
pillow
pandas
numpy
tqdm
opencv-python-headless
scikit-learn